New upstream version 2.8

+++ criu/cr-service.c
@@ -19,7 +19,7 @@
 #include "cr_options.h"
 #include "external.h"
 #include "util.h"
-#include "log.h"
+#include "criu-log.h"
 #include "cpu.h"
 #include "files.h"
 #include "pstree.h"
@@ -304,7 +304,7 @@ static int setup_opts_from_req(int sk, CriuOpts *req)
 	if (req->has_ext_unix_sk) {
 		opts.ext_unix_sk = req->ext_unix_sk;
 		for (i = 0; i < req->n_unix_sk_ino; i++) {
-			if (unix_sk_id_add(req->unix_sk_ino[i]->inode) < 0)
+			if (unix_sk_id_add((unsigned int)req->unix_sk_ino[i]->inode) < 0)
 				goto err;
 		}
 	}
